+++ packages/modal/src/lib/Modal.tsx
@@ -1,0 +1,24 @@
+import { Modal as MuiModal, ModalProps as MuiModalProps } from '@mui/material';
+
+export type ModalProps = {
+  children?: React.ReactNode;
+} & Omit<
+  MuiModalProps,
+  | 'BackdropComponent'
+  | 'BackdropProps'
+  | 'component'
+  | 'components'
+  | 'componentsProps'
+  | 'disableAutoFocus'
+  | 'disableEnforceFocus'
+  | 'disablePortal'
+  | 'onBackdropClick'
+>;
+
+export const Modal = ({ children, ...rest }: ModalProps): JSX.Element => {
+  return (
+    <MuiModal {...rest} disablePortal>
+      {children}
+    </MuiModal>
+  );
+};