+++ packages/checkbox/CHANGELOG.md
@@ -2,6 +2,26 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.1.6](https://github.com/Availity/element/compare/@availity/mui-checkbox@0.1.5...@availity/mui-checkbox@0.1.6) (2023-11-08)
+
+### Dependency Updates
+
+- `mui-form-utils` updated to version `0.5.3`
+- `mui-page-header` updated to version `0.1.6`
+
+## [0.1.5](https://github.com/Availity/element/compare/@availity/mui-checkbox@0.1.4...@availity/mui-checkbox@0.1.5) (2023-11-08)
+
+### Dependency Updates
+
+- `mui-form-utils` updated to version `0.5.2`
+- `mui-icon` updated to version `0.7.2`
+- `mui-typography` updated to version `0.1.4`
+- `mui-page-header` updated to version `0.1.5`
+
+## [0.1.4](https://github.com/Availity/element/compare/@availity/mui-checkbox@0.1.3...@availity/mui-checkbox@0.1.4) (2023-11-08)
+
+## [0.1.3](https://github.com/Availity/element/compare/@availity/mui-checkbox@0.1.2...@availity/mui-checkbox@0.1.3) (2023-11-07)
+
 ## [0.1.2](https://github.com/Availity/element/compare/@availity/mui-checkbox@0.1.1...@availity/mui-checkbox@0.1.2) (2023-10-17)
 
 ## [0.1.1](https://github.com/Availity/element/compare/@availity/mui-checkbox@0.1.0...@availity/mui-checkbox@0.1.1) (2023-10-04)