+++ packages/menu/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.1.6](https://github.com/Availity/element/compare/@availity/mui-menu@0.1.5...@availity/mui-menu@0.1.6) (2023-11-08)
+
+### Dependency Updates
+
+- `mui-button` updated to version `0.5.0`
+
+## [0.1.5](https://github.com/Availity/element/compare/@availity/mui-menu@0.1.4...@availity/mui-menu@0.1.5) (2023-11-08)
+
+### Dependency Updates
+
+- `mui-icon` updated to version `0.7.2`
+- `mui-button` updated to version `0.4.5`
+
+## [0.1.4](https://github.com/Availity/element/compare/@availity/mui-menu@0.1.3...@availity/mui-menu@0.1.4) (2023-11-07)
+
 ## [0.1.3](https://github.com/Availity/element/compare/@availity/mui-menu@0.1.2...@availity/mui-menu@0.1.3) (2023-10-12)
 
 ## [0.1.2](https://github.com/Availity/element/compare/@availity/mui-menu@0.1.1...@availity/mui-menu@0.1.2) (2023-10-06)